refactor(categories): #10 カテゴリ名の一意性制約を廃止
仕様上不要なクエリを削減するため、カテゴリ名の重複チェックを撤廃。
これに伴い CategoryConflictException・findCategoryByName・
buildCreateCategoryWorkflow（ビジネスロジックなし）を削除し、
POST ハンドラではドメインファクトリを route 層で直接呼び出す形に整理。
