+++ docs/spec/functions/02-category.md
@@ -15,7 +15,7 @@
 - 階層構造は持たない（フラット）
 - ユーザによる追加・編集が可能
 - 編集（名称変更）は `status=active` のカテゴリのみ可能
-- カテゴリ名は空文字不可、前後空白除去後に一意
+- カテゴリ名は空文字不可（前後空白除去後）、同名の重複は許可
 - `status` は `active` / `archived` を取る
 - `archived` のカテゴリは新規トランザクション作成・編集時の選択肢に表示しない
 - `archived` のカテゴリへの新規指定・変更を伴うトランザクション登録・更新は業務ルール違反として拒否する
