+++ packages/app/backend/features/categories/exceptions.ts
@@ -5,13 +5,6 @@ export class CategoryNotFoundException extends Error {
   }
 }
 
-export class CategoryConflictException extends Error {
-  constructor(...args: ConstructorParameters<ErrorConstructor>) {
-    super(...args)
-    this.name = 'CategoryConflictException'
-  }
-}
-
 export class CategoryValidationException extends Error {
   constructor(...args: ConstructorParameters<ErrorConstructor>) {
     super(...args)
